[CIR] Support returning and argument floating point types
This adds some of the infrastructure for floating point arg/return
types. The GetSSE type method is stubbed out and just returns the input.
This propagates upwards in a later patch to avoid doing any ABI
concerns at this high of a level.

Keep it around for now, though, for the same reason as always: this
helps us know what's done in clang.