[CIR] Add a hush of a subclass for mlir::OpBuilder for CIR
The llvm::IRBuilder class has some regarding floating point behaviors
that need to be traced for lowering from C++. The behavior largely
mirrors that from the lang options, so it should exist at the highest
level CIR as well.

This patch just builds the hush and replaces the previous
mlir::OpBuilder with the subclass.