[CIR][CodeGen][Lowering] Support Integer overflow with fwrap (#539)
This PR fixes some cases when a program compiled with `-fwrapv` fails
with `NYI` .
Basically, the default behavior  is no overlap:
```
void baz(int x, int y) {
  int z = x - y;
}
```
LLVM IR (no CIR enabled):
```
%sub = sub nsw i32 %0, %1
```
and with `-fwrapv` :
```
%sub = sub i32 %0, %1
```
We need something similar in CIR. The only way I see how to implement it
is to add a couple of attributes to the `BinOp` to make things even with
the llvm dialect.

Well, are there any other ideas?

---------

Co-authored-by: Bruno Cardoso Lopes <bcardosolopes@users.noreply.github.com>