[CIR][CIRGen] Add dynamic builtin alloca intrinsics support (#547)
This patch adds the CIRGen for the following builtin functions:

- `alloca`;
- `_alloca`;
- `__builtin_alloca`;
- `__builtin_alloca_uninitialized`.

Missing support to add in the future:
- Non-default auto initialization setting. The default is to not
initialize the allocated buffer, which is simpler to implement. This
commit is leaving the skeleton to implement this feature following
clang's codegen pattern.
- It may be possible that the frontend has set non-default address space
for the alloca's return value. This is the case for OpenCL or AMDGPU
codes for example. This is handled in clang codegen via address space
cast, and is left for future implementation. This commit introduces a
guard-rail around this behaviour.